add cargo & travel apps
