+++ airport/urls.py
@@ -3,23 +3,13 @@ from rest_framework.routers import DefaultRouter
 
 from airport.views import (
     AirportViewSet,
-    CargoViewSet,
-    CargoAirplaneViewSet,
     PilotViewSet,
-    CargoFlightViewSet,
-    TravelFlightViewSet,
     RouteViewSet,
-    TravelAirplaneViewSet
 )
 
 router = DefaultRouter()
 router.register("airports", AirportViewSet)
-router.register("cargos", CargoViewSet)
-router.register("cargo-airplanes", CargoAirplaneViewSet)
-router.register("travel-airplanes", TravelAirplaneViewSet)
 router.register("pilots", PilotViewSet)
-router.register("cargo-flights", CargoFlightViewSet)
-router.register("travel-flights", TravelFlightViewSet)
 router.register("routes", RouteViewSet)
 
 urlpatterns = [
