+++ config/settings.py
@@ -39,6 +39,8 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "rest_framework",
     "airport",
+    "cargo",
+    "travel",
 ]
 
 MIDDLEWARE = [
