+++ config/urls.py
@@ -20,5 +20,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("api/", include("airport.urls", namespace="airport")),
+    path("api/v1/airport/", include("airport.urls", namespace="airport")),
+    path("api/v1/travel/", include("travel.urls", namespace="travel")),
+    path("api/v1/cargo/", include("cargo.urls", namespace="cargo")),
 ]
